+++ x/erc20/types/params.go
@@ -4,9 +4,11 @@
 package types
 
 import (
+	"bytes"
 	"fmt"
 	"slices"
 
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/evmos/evmos/v19/types"
 )
 
@@ -66,53 +68,83 @@ func (p Params) Validate() error {
 		return err
 	}
 
-	if err := ValidatePrecompiles(p.NativePrecompiles); err != nil {
+	npAddrs, err := ValidatePrecompiles(p.NativePrecompiles)
+	if err != nil {
 		return err
 	}
 
-	if err := ValidatePrecompiles(p.DynamicPrecompiles); err != nil {
+	dpAddrs, err := ValidatePrecompiles(p.DynamicPrecompiles)
+	if err != nil {
 		return err
 	}
 
-	combined := p.DynamicPrecompiles
-	combined = append(combined, p.NativePrecompiles...)
-	return ValidatePrecompilesUniqueness(combined)
+	combined := dpAddrs
+	combined = append(combined, npAddrs...)
+	return validatePrecompilesUniqueness(combined)
 }
 
 // ValidatePrecompiles checks if the precompile addresses are valid and unique.
-func ValidatePrecompiles(i interface{}) error {
+func ValidatePrecompiles(i interface{}) ([]common.Address, error) {
 	precompiles, ok := i.([]string)
 	if !ok {
-		return fmt.Errorf("invalid precompile slice type: %T", i)
+		return nil, fmt.Errorf("invalid precompile slice type: %T", i)
 	}
 
+	precAddrs := make([]common.Address, 0, len(precompiles))
 	for _, precompile := range precompiles {
-		if err := types.ValidateAddress(precompile); err != nil {
-			return fmt.Errorf("invalid precompile %s", precompile)
+		err := types.ValidateAddress(precompile)
+		if err != nil {
+			return nil, fmt.Errorf("invalid precompile %s", precompile)
 		}
+		precAddrs = append(precAddrs, common.HexToAddress(precompile))
 	}
 
 	// NOTE: Check that the precompiles are sorted. This is required
 	// to ensure determinism
 	if !slices.IsSorted(precompiles) {
-		return fmt.Errorf("precompiles need to be sorted: %s", precompiles)
+		return nil, fmt.Errorf("precompiles need to be sorted: %s", precompiles)
 	}
-	return nil
+	return precAddrs, nil
 }
 
-func ValidatePrecompilesUniqueness(i interface{}) error {
-	precompiles, ok := i.([]string)
+func validatePrecompilesUniqueness(i interface{}) error {
+	precompiles, ok := i.([]common.Address)
 	if !ok {
 		return fmt.Errorf("invalid precompile slice type: %T", i)
 	}
 
 	seenPrecompiles := make(map[string]struct{})
 	for _, precompile := range precompiles {
-		if _, ok := seenPrecompiles[precompile]; ok {
+		// use address.Hex() to make sure all addresses are using EIP-55
+		if _, ok := seenPrecompiles[precompile.Hex()]; ok {
 			return fmt.Errorf("duplicate precompile %s", precompile)
 		}
 
-		seenPrecompiles[precompile] = struct{}{}
+		seenPrecompiles[precompile.Hex()] = struct{}{}
 	}
 	return nil
+}
+
+// IsNativePrecompile checks if the provided address is within the native precompiles
+func (p Params) IsNativePrecompile(addr common.Address) bool {
+	return isAddrIncluded(addr, p.NativePrecompiles)
+}
+
+// IsDynamicPrecompile checks if the provided address is within the dynamic precompiles
+func (p Params) IsDynamicPrecompile(addr common.Address) bool {
+	return isAddrIncluded(addr, p.DynamicPrecompiles)
+}
+
+// isAddrIncluded checks if the provided common.Address is within a slice
+// of hex string addresses
+func isAddrIncluded(addr common.Address, strAddrs []string) bool {
+	for _, sa := range strAddrs {
+		// check address bytes instead of the string due to possible differences
+		// on the address string related to EIP-55
+		cmnAddr := common.HexToAddress(sa)
+		if bytes.Equal(addr.Bytes(), cmnAddr.Bytes()) {
+			return true
+		}
+	}
+	return false
 }